requirements.txt
flask==3.0.3
slack-sdk==3.27.2
slack-bolt==1.18.1
anthropic==0.40.0
httpx==0.27.0
databricks-sql-connector==3.3.0
chromadb==0.5.3
sentence-transformers==3.0.1
pandas>=1.2.5,<2.2.0
python-dotenv==1.0.1
